Extract slideshow component from thehowitworks component
